moved from backend -> ops; adjusted the return type
